# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "todos LTDA",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/credit-card": {
      "get": {
        "summary": "List all credit cards",
        "responses": {
          "200": {
            "description": "List of credit cards"
          }
        },
        "tags": ["Credit Cards"]
      },
      "post": {
        "summary": "Create a new credit card",
        "parameters": [
          {
            "in": "body",
            "name": "credit_card_data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "number": {
                  "type": "string"
                },
                "exp_date": {
                  "type": "string",
                  "description": "The expiration date of the credit card (e.g., '12/25')."
                },
                "cvv": {
                  "type": "string",
                  "description": "The CVV code is an optional number."
                },
                "holder": {
                  "type": "string",
                  "description": "The holder is a card name."
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Credit card created successfully"
          },
          "400": {
            "description": "Invalid request data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Failed to create credit card"
          }
        },
        "tags": ["Credit Cards"],
        "security": [
          {
            "Bearer Auth": []
          }
        ]
      }
    },
    "/api/v1/credit-card/{card_id}": {
      "get": {
        "summary": "Get a credit card by ID",
        "parameters": [
          {
            "in": "path",
            "name": "card_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Credit card details"
          },
          "404": {
            "description": "Credit card not found"
          }
        },
        "tags": ["Credit Cards"]
      },
      "put": {
        "summary": "Update a credit card by ID",
        "parameters": [
          {
            "in": "path",
            "name": "card_id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "credit_card_data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "number": {
                  "type": "string"
                },
                "exp_date": {
                  "type": "string",
                  "description": "The expiration date of the credit card (e.g., '12/25')."
                },
                "cvv": {
                  "type": "string",
                  "description": "The CVV code is an optional number."
                },
                "holder": {
                  "type": "string",
                  "description": "The holder is a card name."
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Credit card updated successfully"
          },
          "400": {
            "description": "Invalid request data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Credit card not found"
          },
          "500": {
            "description": "Failed to update credit card"
          }
        },
        "tags": ["Credit Cards"],
        "security": [
          {
            "Bearer Auth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete a credit card by ID",
        "parameters": [
          {
            "in": "path",
            "name": "card_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "Credit card deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Credit card not found"
          },
          "500": {
            "description": "Failed to delete credit card"
          }
        },
        "tags": ["Credit Cards"],
        "security": [
          {
            "Bearer Auth": []
          }
        ]
      }
    },
    "/api/v1/register": {
      "post": {
        "summary": "Register a new user",
        "parameters": [
          {
            "in": "body",
            "name": "user_data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "confirm_password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Invalid request data"
          },
          "500": {
            "description": "Failed to register user"
          }
        },
        "tags": ["User"]
      }
    },
    "/api/v1/login": {
      "post": {
        "summary": "User login",
        "parameters": [
          {
            "in": "body",
            "name": "user_data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User logged in successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Invalid request data"
          }
        },
        "tags": ["User"]
      }
    },
    "/api/v1/user/{username}": {
      "get": {
        "summary": "Get user details by email",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User details"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": ["User"]
      }
    }
  },
  "tags": [
    {
      "name": "Credit Cards",
      "description": "Operations related to credit cards"
    },
    {
      "name": "User",
      "description": "Operations related to users"
    }
  ],
  "securityDefinitions": {
    "Bearer Auth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}